tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/animation-name_one_value.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - animation-name value with one list entry</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#the-animation-name-property-">
   <meta name="flags" content="dom">
   <meta name="assert" content="IDENT is a valid animation-name value">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes fade {
         from {
            opacity: 1;
         }
         to {
            opacity: 0;
         }
      }
      div {
         width: 100px;
         height: 100px;
         background: blue;
      }
      .test {
         animation-duration: 1s;
         animation-name: fade;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var div = document.querySelector('div');
         var log = document.querySelectorAll('p')[1].firstChild;
         var timeout = setTimeout(function() {
            test(function(){assert_true(false);}, testName); done();
            log.data = 'FAIL';
         }, 1500);
         div.addEventListener('animationend', function(evt) {
            clearTimeout(timeout);
            var passed = (evt.animationName == 'fade');
            test(function(){assert_true(passed);}, testName); done();
            log.data = (passed ? 'PASS if the animation occurred.' : 'FAIL');
         });
         div.className = 'test';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>The block below should fade out and then reappear.</p>
   <p>Testing...</p>
   <div></div>
</body>
</html>
